Added To-do list and more VS2010 stuff.
I'd hate to mess with the engine when OpenJK can do it so much better, but the weapon system as is angers me in ways I can't explain.
Switching over to C++ and making it more object oriented will be much nicer.
Not a full port, just the weapon system so we can add new weapons with one class and a few edits.